+++ components/Sales/Direct/Button.tsx
@@ -1,7 +1,14 @@
-import { Button, Flex, Text } from '@chakra-ui/react'
+import { Button, Flex, Text, useDisclosure, useToast } from '@chakra-ui/react'
+import { CancelOfferStep, useCancelOffer } from '@liteflow/react'
 import { HiArrowNarrowRight } from '@react-icons/all-files/hi/HiArrowNarrowRight'
+import ConnectButtonWithNetworkSwitch from 'components/Button/ConnectWithNetworkSwitch'
+import CancelOfferModal from 'components/Modal/CancelOffer'
 import useTranslation from 'next-translate/useTranslation'
-import { FC, useMemo } from 'react'
+import { FC, useCallback, useMemo } from 'react'
+import { useAccount } from 'wagmi'
+import useBlockExplorer from '../../../hooks/useBlockExplorer'
+import useSigner from '../../../hooks/useSigner'
+import { formatError, isSameAddress } from '../../../utils'
 import Link from '../../Link/Link'
 import type { Props as ModalProps } from './Modal'
 import SaleDirectModal from './Modal'
@@ -24,6 +31,76 @@ const SaleDirectButton: FC<Props> = ({
   onOfferCanceled,
 }) => {
   const { t } = useTranslation('components')
+  const signer = useSigner()
+  const [cancelOffer, { activeStep, transactionHash }] = useCancelOffer(signer)
+  const toast = useToast()
+  const { address } = useAccount()
+  const blockExplorer = useBlockExplorer(chainId)
+  const { isOpen, onOpen, onClose } = useDisclosure()
+
+  const handleCancel = useCallback(
+    async (saleId: string) => {
+      if (!confirm(t('sales.direct.info.cancel-confirmation'))) return
+      try {
+        onOpen()
+        await cancelOffer(saleId)
+        await onOfferCanceled(saleId)
+      } catch (e) {
+        toast({
+          title: formatError(e),
+          status: 'error',
+        })
+      } finally {
+        onClose()
+      }
+    },
+    [t, onOpen, cancelOffer, onOfferCanceled, toast, onClose],
+  )
+
+  const cancel = useMemo(() => {
+    if (sales.length !== 1) return
+    // Only display cancel when there is a single offer owned by the current account
+    if (!address) return
+    const sale = sales[0]
+    if (!sale) return
+    const currentAccountFirstSale = isSameAddress(sale.maker.address, address)
+    if (!currentAccountFirstSale) return
+    return (
+      <>
+        <ConnectButtonWithNetworkSwitch
+          chainId={chainId}
+          onClick={() => handleCancel(sale.id)}
+          isLoading={activeStep !== CancelOfferStep.INITIAL}
+          width="full"
+          size="lg"
+        >
+          <Text as="span" isTruncated>
+            {t('sales.direct.info.cancel')}
+          </Text>
+        </ConnectButtonWithNetworkSwitch>
+        <CancelOfferModal
+          isOpen={isOpen}
+          onClose={onClose}
+          title={t('sales.direct.info.cancel')}
+          step={activeStep}
+          blockExplorer={blockExplorer}
+          transactionHash={transactionHash}
+        />
+      </>
+    )
+  }, [
+    activeStep,
+    address,
+    blockExplorer,
+    chainId,
+    handleCancel,
+    isOpen,
+    onClose,
+    sales,
+    t,
+    transactionHash,
+  ])
+
   const bid = useMemo(() => {
     if (ownAllSupply) return
     return (
@@ -44,21 +121,19 @@ const SaleDirectButton: FC<Props> = ({
 
   const buyNow = useMemo(() => {
     if (sales.length !== 1) return
-    if (!sales[0]) return
-    if (ownAllSupply) return
+    if (!address) return
+    const sale = sales[0]
+    if (!sale) return
+    // Only display buy now when there is a single offer not owned by the current account
+    if (isSameAddress(sale.maker.address, address)) return
     return (
-      <Button
-        as={Link}
-        href={`/checkout/${sales[0].id}`}
-        size="lg"
-        width="full"
-      >
+      <Button as={Link} href={`/checkout/${sale.id}`} size="lg" width="full">
         <Text as="span" isTruncated>
           {t('sales.direct.button.buy')}
         </Text>
       </Button>
     )
-  }, [sales, ownAllSupply, t])
+  }, [address, sales, t])
 
   const seeOffers = useMemo(() => {
     if (sales.length <= 1) return
@@ -88,11 +163,12 @@ const SaleDirectButton: FC<Props> = ({
       </Button>
     )
 
-  if (!bid && !buyNow && !seeOffers) return null
+  if (!cancel && !bid && !buyNow && !seeOffers) return null
 
   return (
     <Flex gap={6} direction={{ base: 'column', md: 'row' }}>
       {bid}
+      {cancel}
       {buyNow}
       {seeOffers}
     </Flex>
